third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/stroke-miterlimit-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'stroke-miterlimit' to a number: -3.14
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
Harness: the test ran to completion.
